+++ kparser2.core/AnalyticsDtoMapping.fs
@@ -111,9 +111,11 @@ module AnalyticsDtoMapping =
             MessageId = i.MessageId,
             IsProc = i.IsProc,
             ProcValue = i.ProcValue,
+            ProcMessageId = i.ProcMessageId,
             IsLocalPlayerActor = i.IsLocalPlayerActor,
             IsLocalPlayerTarget = i.IsLocalPlayerTarget,
-            SourcePacketId = Option.toObj i.SourcePacketId
+            SourcePacketId = Option.toObj i.SourcePacketId,
+            IsHpDamage = InteractionClassification.isHpDamage i
         )
 
     let private toChat (c: ChatMessageRecord) =
@@ -239,6 +241,7 @@ module AnalyticsDtoMapping =
           MessageId = i.MessageId
           IsProc = i.IsProc
           ProcValue = i.ProcValue
+          ProcMessageId = i.ProcMessageId
           IsLocalPlayerActor = i.IsLocalPlayerActor
           IsLocalPlayerTarget = i.IsLocalPlayerTarget
           SourcePacketId = Option.ofObj i.SourcePacketId }